refactor: 사이드 바의 Suspense를 토픽 불러오는 부분으로 이동
